+++ lihil/config/app_config.py
@@ -8,34 +8,34 @@ from lihil.interface import Record
 
 class IOASConfig(Protocol):
     @property
-    def oas_path(self) -> str: ...
+    def OAS_PATH(self) -> str: ...
     @property
-    def doc_path(self) -> str: ...
+    def DOC_PATH(self) -> str: ...
     @property
-    def title(self) -> str: ...
+    def TITLE(self) -> str: ...
     @property
-    def problem_path(self) -> str: ...
+    def PROBLEM_PATH(self) -> str: ...
     @property
-    def problem_title(self) -> str: ...
+    def PROBLEM_TITLE(self) -> str: ...
     @property
-    def version(self) -> str: ...
+    def VERSION(self) -> str: ...
 
 
 class IServerConfig(Protocol):
     @property
-    def host(self) -> str: ...
+    def HOST(self) -> str: ...
     @property
-    def port(self) -> int: ...
+    def PORT(self) -> int: ...
     @property
-    def workers(self) -> int: ...
+    def WORKERS(self) -> int: ...
     @property
-    def reload(self) -> bool: ...
+    def RELOAD(self) -> bool: ...
     def asdict(self) -> dict[str, Any]: ...
 
 
 class IAppConfig(Protocol):
     @property
-    def version(self) -> str: ...
+    def VERSION(self) -> str: ...
     @property
     def server(self) -> IServerConfig: ...
     @property
@@ -46,33 +46,33 @@ class ConfigBase(Record, forbid_unknown_fields=True): ...
 
 
 class OASConfig(ConfigBase):
-    oas_path: Annotated[str, Doc("Route path for OpenAPI JSON schema")] = "/openapi"
-    doc_path: Annotated[str, Doc("Route path for Swagger UI")] = "/docs"
-    title: Annotated[str, Doc("Title of your Swagger UI")] = "lihil-OpenAPI"
-    problem_path: Annotated[str, Doc("Route path for problem page")] = "/problems"
-    problem_title: Annotated[str, Doc("Title of your problem page")] = (
+    OAS_PATH: Annotated[str, Doc("Route path for OpenAPI JSON schema")] = "/openapi"
+    DOC_PATH: Annotated[str, Doc("Route path for Swagger UI")] = "/docs"
+    TITLE: Annotated[str, Doc("Title of your Swagger UI")] = "lihil-OpenAPI"
+    PROBLEM_PATH: Annotated[str, Doc("Route path for problem page")] = "/problems"
+    PROBLEM_TITLE: Annotated[str, Doc("Title of your problem page")] = (
         "lihil-Problem Page"
     )
-    version: Annotated[str, Doc("Swagger UI version")] = "3.1.0"
+    VERSION: Annotated[str, Doc("Swagger UI version")] = "3.1.0"
 
 
 class ServerConfig(ConfigBase):
-    host: Annotated[str, Doc("Host address to bind to (e.g., '127.0.0.1')")] = (
+    HOST: Annotated[str, Doc("Host address to bind to (e.g., '127.0.0.1')")] = (
         "127.0.0.1"
     )
-    port: Annotated[int, Doc("Port number to listen on e.g., 8000")] = 8000
-    workers: Annotated[int, Doc("Number of worker processes")] = 1
-    reload: Annotated[bool, Doc("Enable auto-reloading during development")] = False
-    root_path: Annotated[
+    PORT: Annotated[int, Doc("Port number to listen on e.g., 8000")] = 8000
+    WORKERS: Annotated[int, Doc("Number of worker processes")] = 1
+    RELOAD: Annotated[bool, Doc("Enable auto-reloading during development")] = False
+    ROOT_PATH: Annotated[
         str, Doc("Root path to mount the app under (if behind a proxy)")
     ] = ""
 
 
 class AppConfig(ConfigBase):
-    is_prod: Annotated[bool, Doc("Whether the current environment is production")] = (
+    IS_PROD: Annotated[bool, Doc("Whether the current environment is production")] = (
         False
     )
-    version: Annotated[str, Doc("Application version")] = "0.1.0"
+    VERSION: Annotated[str, Doc("Application version")] = "0.1.0"
     oas: Annotated[OASConfig, Doc("OpenAPI and Swagger UI configuration")] = field(
         default_factory=OASConfig
     )