+++ lihil/plugins/__init__.py
@@ -4,33 +4,32 @@ mostly simple wrappers to third-party dependencies.
 if not, likely to be a standalone lib
 """
 
-from typing import Any, Awaitable, Callable, Protocol
+from typing import Any, Protocol, Generic
 
 from ididi import Graph
 
+from lihil.interface import IAsyncFunc, P, R
 from lihil.signature import EndpointSignature
 
-IFunc = Callable[..., Awaitable[Any]]
 
-
-class IAsyncPlugin(Protocol):
+class IAsyncPlugin(Protocol, Generic[P, R]):
     async def __call__(
         self,
         graph: Graph,
-        func: IFunc,
+        func: IAsyncFunc[P, R],
         sig: EndpointSignature[Any],
         /,
-    ) -> IFunc: ...
+    ) -> IAsyncFunc[P, R]: ...
 
 
-class ISyncPlugin(Protocol):
+class ISyncPlugin(Protocol, Generic[P,R]):
     def __call__(
         self,
         graph: Graph,
-        func: IFunc,
+        func: IAsyncFunc[P, R],
         sig: EndpointSignature[Any],
         /,
-    ) -> IFunc: ...
+    ) -> IAsyncFunc[P, R]: ...
 
 
-IPlugin = IAsyncPlugin | ISyncPlugin
+IPlugin = IAsyncPlugin[..., Any] | ISyncPlugin[..., Any]