+++ lihil/plugins/auth/supabase.py
@@ -15,14 +15,14 @@ from gotrue.types import SignInWithIdTokenCredentials as SignInWithIdTokenCreden
 
 
 
-class SupabasePlugin:
-    def __init__(self, supabase: AsyncClient):
-        self.supabase = supabase
+# class SupabasePlugin:
+#     def __init__(self, supabase: AsyncClient):
+#         self.supabase = supabase
 
-    async def __call__(self, email: str, password: str):
-        ...
+#     async def __call__(self, email: str, password: str):
+#         ...
 
-        self.supabase.auth.sign_in_with_id_token()
+#         self.supabase.auth.sign_in_with_id_token()
         # user = await self.supabase.auth.sign_up(email=email, password=password)
         # return user