+++ lihil/signature/returns.py
@@ -108,10 +108,6 @@ def parse_return_pro(
                 content_type = metas[idx + 1]
         elif is_status(meta):
             status = get_status_code(meta)
-        elif isinstance(meta, ParamMeta) and meta.extra and meta.extra.use_jwt:
-            from lihil.plugins.auth.jwt import jwt_encoder_factory
-
-            encoder = jwt_encoder_factory(payload_type=ret_type)
         else:
             continue